Moved creators out of Factory
