+++ lib/UnitFactory.js
@@ -11,6 +11,28 @@ UnitFactory.createUnit = function(unitType, location) {
     var stats;
     // MAKE SURE YOU ADD TO UnitType and Player Factor too!
     switch(unitType) {
+        case UnitType.AMBROSE:
+            stats = {
+                "name": "ambrose",
+                "description": "A lvl 1 noob",
+                "hint": "Kill him with fire!!",
+                "team": Team.CREEP,
+                "type": unitType,
+                "maxHp": 10,
+                "location": location,
+                "movementDelay": 1,
+                "meleeAttackPower": 1,
+                "meleeAttackDelay": 1,
+                "magicAttackPower": 0,
+                "magicAttackDelay": 0,
+                "rangedAttackPower": 0,
+                "rangedAttackDelay": 0,
+                "sensorRadiusSquared": 49,
+                "alertRadiusSquared": 16,
+                "spawnDelay": 0,
+                "spawnedUnitType": null
+            };
+            return new Unit(stats);
         case UnitType.GNAT:
             stats = {
                 "name": "gnat",
