+++ lib/UnitType.js
@@ -1,4 +1,5 @@
 module.exports = {
+    AMBROSE: 'a',
     GNAT: 'g',
     NECROMANCER: 'n',
     QUILL_BOAR: 'q',
